Add files via upload
Problem Statement: 
In India, the industrial classification of the workforce is essential to understand the distribution of the labor force across various sectors. The classification of main workers and marginal workers, other than cultivators and agricultural laborers, by sex and by section, division, and class, has been traditionally used to understand the economic status and employment trends in the country. However, the current data on this classification is outdated and may not accurately reflect the current state of the workforce. The aim of this study is to update the information on the industrial classification of the main and marginal workers, other than cultivators and agricultural laborers, by sex and by section, division, and class, to provide relevant and accurate data for policy making and employment planning

Approach:
1) Merge all the csv data file provided to you and create dataframe
2) The classical machine learning tasks like Data Exploration, Data Cleaning, Feature Engineering, Model Building and Model Testing. Use Natural Language Processing for analyzing the various core industries and group the business categories like Retail, Poultry, Agriculture, Manufacturing etc